Replaces tabs with spaces.
To keep code base consistent. client.py was in indented with tabs and setup.py with spaces.
The python style guide (https://www.python.org/dev/peps/pep-0008/#tabs-or-spaces) says:
'Spaces are the preferred indentation method.
Tabs should be used solely to remain consistent with code that is already indented with tabs.
Python 3 disallows mixing the use of tabs and spaces for indentation.
Python 2 code indented with a mixture of tabs and spaces should be converted to using spaces exclusively.'